[#2482] Move find_signal_by_name to utils
